first 50 Fibonacci numbers, starting with 1 and 2 ,fixed

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	long int y[50] = {1, 2};
+	double y[50] = {1, 2};
 	int count;
 
 	printf("1, 2, ");
 	for (count = 2 ; count < 51 ; count++)
 	{
 		y[count] = y[count - 1] + y[count - 2];
-		printf("%ld", y[count]);
+		printf("%0.1f", y[count]);
 		if (count < 49)
 			printf(", ");
 	}
